+++ src/pk/mod.rs
@@ -1,5 +1,6 @@
 pub mod absorption;
 pub mod event_driven;
+pub mod ode_template;
 pub mod one_compartment;
 pub mod three_compartment;
 pub mod two_compartment;